keep available variables, skipping non-available ones
Suppose that you have many datasets. You want to keep a list of variables in each dataset. However, some variables may not be available in some datasets. In that case, if you use the "keep [varlist]" command, it will cause an error when a dataset does not contain a certain variable in varlist. That error will also terminate the rest of the do file. This do file shows how to keep available variables, automatically skipping non-available ones.